Move Edge struct to own file to conform to Style Guide
